+++ flutter_lmdb2/lib/flutter_lmdb2.dart
@@ -1,1 +1,0 @@
-export 'package:dart_lmdb2/dart_lmdb2.dart';