+++ flutter_lmdb2/lib/lmdb.dart
@@ -1,0 +1,1 @@
+export 'package:dart_lmdb2/lmdb.dart';